修正-rpop入参

+++ armagin-server/src/main/java/cc/forim/armagin/server/infra/utils/RedisUtil.java
@@ -48,7 +48,7 @@ public final class RedisUtil {
     /**
      * 右弹出size个元素
      */
-    public List<?> rPop(String key, Integer size) {
+    public List<?> rPop(String key, Long size) {
         return redisTemplate.opsForList().rightPop(key, size);
     }
